* Align/Aligner.h (constructor) <buckets>: size_t not int.
git-svn-id: https://svn.bcgsc.ca/svn/abyss/trunk@52718 ad02020f-2a26-4133-947e-1829317c1416

+++ Align/Aligner.h
@@ -148,10 +148,10 @@ class Aligner
 		typedef typename SeqPosHashMap::const_iterator
 			map_const_iterator;
 
-		Aligner(int hashSize, int buckets)
+		Aligner(int hashSize, size_t buckets)
 			: m_hashSize(hashSize), m_target(buckets) { }
 
-		Aligner(int hashSize, int buckets, float factor)
+		Aligner(int hashSize, size_t buckets, float factor)
 			: m_hashSize(hashSize)
 		{
 			m_target.max_load_factor(factor);
